refactor(mobile): damai_app.py 拆分为包（零行为变更）
把单文件 mobile/damai_app.py（2072 行）拆分为 mobile/damai_app/ 包：
- __init__.py        包入口；保留所有原模块级符号 + 自定义 ModuleType
                     hook 把 patch("mobile.damai_app.{time,datetime,logger}")
                     镜像到所有子模块，使测试零修改
- __main__.py        支持 python -m damai_app（替代原 python damai_app.py）
- orchestrator.py    DamaiBot 主类 + run_ticket_grabbing + run_with_retry
                     + 启动/弹窗/会话探测/票价确认等编排
- sale_waiter.py     wait_for_sale_start + 自适应轮询
- purchase_flow.py   _enter_purchase_flow_from_detail_page + _submit_order_fast
- recovery_strategies.py 状态机 dispatcher + back-press 恢复
- coords_cache.py    热路径坐标缓存命中检查与坐标读取
- state_probe.py     页面状态探测 + SKU 检视 + 标题/场地解析
- delegators.py      AttendeeSelector / PriceSelector / EventNavigator /
                     FastPipeline 的薄封装

零行为变更：
- 仅搬代码，不动逻辑、不改默认值、不重命名（_submit_order_fast 等保持下划线）
- 外部 API 不变：from mobile.damai_app import DamaiBot 仍兼容；
  Config / logger / time / datetime 等模块属性仍可被测试 patch
- 测试零修改：1020 通过 + 1 预存 xfail，覆盖率 81.03%（≥80% 门槛）

启动脚本同步：
- mobile/scripts/start_ticket_grabbing.sh：python damai_app.py → python -m damai_app
